+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 v4.0
 ====
 
-- [feat] add `valueFormat` and `valueFormatType` properties to separately control real value and visuals (property `format`)
-- [breaking] `format` and `formatType` now controls only visual representation of selected date time. Value format defaults to `yyyy-mm-dd` 
+- [feat] add `displayFormat` and `displayFormatType` properties to separately control visuals representation and real value (properties `format`, `formatType`)
 - [breaking] removed `positionFn` property
 - [breaking] removed `clearToggle` property (now this functionality - ability to clear the value is controlled by `required` property)
 - [breaking] method to register component as custom element has changed from `registerElement` to `registerSveltyPicker`. Also arguments has changed